Catching up to master

+++ holberton.h
@@ -7,7 +7,7 @@
 int _putchar(char c, int *k);
 void _puts(char *str, int *k);
 int _printf(const char *format, ...);
-void _print_number(int n, int *k);
+void _print_number(long int n, int *k);
 void _rev_string(char *s, int *k);
 int _strlen(char *s);
 void _print_binary(int n, int *k);
